+++ contracts/vm/ops/token/IERC721Ops.sol
@@ -24,7 +24,7 @@ library IERC721Ops {
             require(opcode_ < IERC721_OPS_LENGTH, "MAX_OPCODE");
 
             state_.stackIndex--;
-            uint baseIndex_ = state_.stackIndex - 1;
+            uint256 baseIndex_ = state_.stackIndex - 1;
             // Stack the return of `balanceOf`.
             if (opcode_ == OPCODE_BALANCE_OF) {
                 state_.stack[baseIndex_] = IERC721(
@@ -35,11 +35,8 @@ library IERC721Ops {
             else if (opcode_ == OPCODE_OWNER_OF) {
                 state_.stack[baseIndex_] = uint256(
                     uint160(
-                        IERC721(
-                            address(
-                                uint160(state_.stack[baseIndex_])
-                            )
-                        ).ownerOf(state_.stack[state_.stackIndex])
+                        IERC721(address(uint160(state_.stack[baseIndex_])))
+                            .ownerOf(state_.stack[state_.stackIndex])
                     )
                 );
             }
